[refactor] custom hook 제작 modal/statchange #775
